Fix tensor_spec_utils.set_bounds() to work correctly with int8 types.
Previously we would convert the min/max args to a numpy array with an inferred type. This would cause issues when later packing the value to a proto field with a smaller type and then unpacking later. For example, packing an int32 into an int8 field would result in 4 bytes being packed instead of 1.

As we now explicitly set the dtype, this will truncate invalid values, so we move the can_cast test to before the conversion.

PiperOrigin-RevId: 344438293
Change-Id: I4e0cb9ce3a426e4379705c387759a99d673158bd