Move exception handling for packing unrecognized types.
Previously we would attempt to raise a helpful error message at the earliest possible moment when attempting to pack an unrecognized value into a Tensor message. This logic was becoming increasingly complex, so instead we raise this error at a later point. This has the added benefit of being able to assert that all elements are indeed proto messages rather than just the first one.

PiperOrigin-RevId: 364847226
Change-Id: Ifa680e97a144bda7344ffad2280764098759c6fa