Allow unsafe casting when attempting to pack an empty array as a particular dtype.
By default, when calling np.asarray() with an empty array, NumPy defaults to dtype float64. This causes an unexpected error when passing an explicit dtype to `pack_tensor()`. For example:

`tensor_utils.pack_tensor([], dtype=np.int64)`

raises a TypeError when trying to cast from float64 to int64. In this particular case, we allow unsafe casting.

PiperOrigin-RevId: 420717157
Change-Id: I6bb8a88e9d0cb246681590a1dbe4488bddb3604b