Change OrderedDict to dict in flatten utils
Ordered dict was used to guarantee insertion order in Python version <=3.6.
As of Python version >=3.7, dict guarantees insertion order as well, so we can
safely drop OrderedDict.

Note that (key, value) order in mapping was not guaranteed in the first place
as data is serialised using proto maps, see
https://developers.google.com/protocol-buffers/docs/proto3#maps.
As a consequence, this is a safe change as downstream project could hardly have
relied on any order from the returned OrderedDict.

PiperOrigin-RevId: 440035637
Change-Id: Ia7bad045ec06829fb832f44a3a47d472fce6189b